+++ handlers/entitlement_handlers.go
@@ -6,7 +6,7 @@ import (
 	"github.com/disgoorg/disgo/gateway"
 )
 
-func gatewayHandlerEntitlementCreate(client bot.Client, sequenceNumber int, shardID int, event gateway.EventEntitlementCreate) {
+func gatewayHandlerEntitlementCreate(client bot.Client, sequenceNumber uint64, shardID int, event gateway.EventEntitlementCreate) {
 	client.EventManager().DispatchEvent(&events.EntitlementCreate{
 		GenericEntitlementEvent: &events.GenericEntitlementEvent{
 			GenericEvent: events.NewGenericEvent(client, sequenceNumber, shardID),
@@ -15,7 +15,7 @@ func gatewayHandlerEntitlementCreate(client bot.Client, sequenceNumber int, shar
 	})
 }
 
-func gatewayHandlerEntitlementUpdate(client bot.Client, sequenceNumber int, shardID int, event gateway.EventEntitlementUpdate) {
+func gatewayHandlerEntitlementUpdate(client bot.Client, sequenceNumber uint64, shardID int, event gateway.EventEntitlementUpdate) {
 	client.EventManager().DispatchEvent(&events.EntitlementUpdate{
 		GenericEntitlementEvent: &events.GenericEntitlementEvent{
 			GenericEvent: events.NewGenericEvent(client, sequenceNumber, shardID),
@@ -24,7 +24,7 @@ func gatewayHandlerEntitlementUpdate(client bot.Client, sequenceNumber int, shar
 	})
 }
 
-func gatewayHandlerEntitlementDelete(client bot.Client, sequenceNumber int, shardID int, event gateway.EventEntitlementDelete) {
+func gatewayHandlerEntitlementDelete(client bot.Client, sequenceNumber uint64, shardID int, event gateway.EventEntitlementDelete) {
 	client.EventManager().DispatchEvent(&events.EntitlementDelete{
 		GenericEntitlementEvent: &events.GenericEntitlementEvent{
 			GenericEvent: events.NewGenericEvent(client, sequenceNumber, shardID),
