+++ handlers/guild_integrations_update_handler.go
@@ -6,7 +6,7 @@ import (
 	"github.com/disgoorg/disgo/gateway"
 )
 
-func gatewayHandlerGuildIntegrationsUpdate(client bot.Client, sequenceNumber int, shardID int, event gateway.EventGuildIntegrationsUpdate) {
+func gatewayHandlerGuildIntegrationsUpdate(client bot.Client, sequenceNumber uint64, shardID int, event gateway.EventGuildIntegrationsUpdate) {
 	client.EventManager().DispatchEvent(&events.GuildIntegrationsUpdate{
 		GenericEvent: events.NewGenericEvent(client, sequenceNumber, shardID),
 		GuildID:      event.GuildID,
