+++ handlers/invite_handlers.go
@@ -6,14 +6,14 @@ import (
 	"github.com/disgoorg/disgo/gateway"
 )
 
-func gatewayHandlerInviteCreate(client bot.Client, sequenceNumber int, shardID int, event gateway.EventInviteCreate) {
+func gatewayHandlerInviteCreate(client bot.Client, sequenceNumber uint64, shardID int, event gateway.EventInviteCreate) {
 	client.EventManager().DispatchEvent(&events.InviteCreate{
 		GenericEvent:      events.NewGenericEvent(client, sequenceNumber, shardID),
 		EventInviteCreate: event,
 	})
 }
 
-func gatewayHandlerInviteDelete(client bot.Client, sequenceNumber int, shardID int, event gateway.EventInviteDelete) {
+func gatewayHandlerInviteDelete(client bot.Client, sequenceNumber uint64, shardID int, event gateway.EventInviteDelete) {
 	client.EventManager().DispatchEvent(&events.InviteDelete{
 		GenericEvent: events.NewGenericEvent(client, sequenceNumber, shardID),
 		GuildID:      event.GuildID,
